src/agent/requirements.txt
google-adk>=1.20.0
litellm>=1.50.0
pydantic>=2.0.0
httpx>=0.25.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
